+++ src/mindlessgen/prog/__init__.py
@@ -1,5 +1,6 @@
 """
-This module contains the classes and functions for all configuration-related tasks.
+This module contains the classes and functions for all configuration-related tasks,
+as well as utilities concerned with parallelization.
 """
 
 from .config import (
@@ -14,6 +15,7 @@ from .config import (
     PostProcessConfig,
     StructureModConfig,
 )
+from .parallel import setup_managers, ResourceMonitor, setup_blocks
 
 __all__ = [
     "ConfigManager",
@@ -25,5 +27,8 @@ __all__ = [
     "GenerateConfig",
     "RefineConfig",
     "PostProcessConfig",
+    "setup_managers",
+    "ResourceMonitor",
+    "setup_blocks",
     "StructureModConfig",
 ]